refactor: move bitfields into seperate folder
