+++ src/util/bitfields/Bitfield.ts
